# orion-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Orion DAG orchestration platform",
        "title": "Orion API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/dags/{dagId}/runs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "List runs for a DAG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG ID",
                        "name": "dagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Run"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Trigger a DAG run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG ID",
                        "name": "dagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Run"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dags/{dagId}/task-instances": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-instances"
                ],
                "summary": "List task instances",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG ID",
                        "name": "dagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TaskInstance"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/runs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Get a run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Run"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/task-instances/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task-instances"
                ],
                "summary": "Get a task instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaskInstance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Run": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dag_id": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "triggered_by": {
                    "type": "string"
                }
            }
        },
...
        "models.TaskInstance": {
            "type": "object",
            "properties": {
                "attempt": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "log_location": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
    }
}